Make C plugin output format configurable
The development guidelines¹ currently state that the plugin output
should be in the format "SERVICE STATUS: Information text".  However,
the SERVICE and STATUS information is redundant, as the plugin caller
(e.g., Nagios) should know which SERVICE is checked and the STATUS is
returned via the plugin's exit code.  After discussing² this issue on
the development list, we decided to make the plugin output configurable
at compile time (separately for "normal" and "verbose" output).

¹ http://nagiosplug.sf.net/developer-guidelines.html
² http://thread.gmane.org/gmane.network.nagios.plugins.devel/5155

Signed-off-by: Holger Weiss <hweiss@users.sourceforge.net>