Make C plugin output format configurable
*** THIS COMMIT WILL BE MODIFIED IN THE FUTURE! ***

The development guidelines¹ currently state that the plugin output
should be in the format "SERVICE STATUS: Information text".

However, when a plugin is called via Nagios in order to perform a
service check, adding SERVICE and STATUS to the output is redundant.
And when a plugin is called on the command line for testing, there is no
use in printing out the SERVICE string, either.  However, for debugging,
it does make sense to print out the STATUS so that the user won't have
to check and interpret the exit code manually.  But it should suffice to
print such debug output only when called with "--verbose", not in
production.

Space for plugin output is sometimes scarce, so that we should try to
keep the output "short and to the point" (as the guide states) and not
waste space with redundant information.  Therefore, we decided² to make
the ("normal" and "verbose") plugin output configurable at compile time.

¹ http://nagiosplug.sf.net/developer-guidelines.html
² http://thread.gmane.org/gmane.network.nagios.plugins.devel/5155