check_http: Support HTTP CONNECT method
Add the -Y/--http-connect[=<port>] option which tells check_http to use
the HTTP CONNECT method for tunneling an HTTPS connection through a
proxy server.

This is a modified version of a patch provided by Mark Frost in
SourceForge tracker item #2975393, updated for the current check_http
code.  Changes include:

- Let the new --http-connect option imply --ssl.
- Allow for specifying the server port the proxy should connect to.