utils.sh: Don't define $ECHO
In utils.sh, $ECHO was set to printf(1) or echo(1), depending on whether
printf(1) is available in /usr/bin.  This resulted in various bugs, as
printf(1) cannot be used in the same way as echo(1).

Thanks to Trevor Hemsley for reporting one of those bugs: If $ECHO was
set to printf(1), check_log stumbled when reporting a log line with a
"%" character.