check_dig: patch to make dig honor -t option
When a timeout value is specified with the -t option, dig will sometimes
timeout before the timer is actually reached.

The problem occurs because the check_dig plugin does not pass the specified
timeout value to dig, leaving dig to timeout with it's default value which
seems to be around 10-15seconds.

To reproduce:
time ./check_dig -H 127.0.0.2 -l www.google.com -t 30

It will not run for 30secs, which is the expected behaviour.

The following will work, because the timeout is less than the default dig
timeout, so the plugin cancels the dig command:
time ./check_dig -H 127.0.0.2 -l www.google.com -t 2

This fix passes the timeout value to dig, and sets the number of retries which tends to vary from system to system by default.

Closes #1168