use FindBin instead of awk to find the path to utils.pm
'use lib utils.pm' is not valid Perl syntax:
 Bareword "utils" not allowed while "strict subs" in use at plugins-scripts/check_ircd.pl line 52.
 Bareword "pm" not allowed while "strict subs" in use at plugins-scripts/check_ircd.pl line 52.

This makes it impossible to use the plugins directly from the git tree,
e.g. while hacking on them.

Using FindBin::Bin as the library path allows that, while preserving
the original behaviour of adding the libexec path when the plugin is
properly installed.