check_swap: Fix perfdata und thresholds for big values and simplify code
The original problem was https://github.com/monitoring-plugins/monitoring-plugins/pull/1705
where the performance data output of check_swap did not conform to
the parser logic of a monitoring system (which decided to go for
"correct" SI or IEC units.
The PR was accompanied by a change to byte values in the performance
data which broke the _perfdata_ helper function which could not handle
values of this size.
The fix for this, was to use _fperfdata_ which could, but would
use float values.

I didn't like that (since all values here are discreet) and this
is my proposal for a fix for the problem.

It introduces some helper functions which do now explicitely work
with (u)int64_t, including a special version of the _perfdata_ helper.

In the process of introducing this to check_swap, I stumbled over
several sections of the check_swap code which I found problematic.
Therefore I tried to simplify the code and make it more readable
and less redundant.

I am kinda sorry about this, but sincerely hope my changes can
be helpful.