configure.ac: require snmpgetnext for check_snmp
PATH_TO_SNMPGETNEXT is used unconditionally in plugins/check_snmp.c,
and the build will fail if it is left undefined (that is, if we are
building check_snmp but snmpgetnext was neither found on the user's
PATH or supplied manually).

To avoid this build failure, we now test for snmpgetnext inside the
case for snmpget, and skip check_snmp unless BOTH are found.